+++ src/pokeclicker/model/item/Item.java
@@ -5,11 +5,11 @@ import pokeclicker.model.common.Purchasable;
 
 public abstract class Item implements Purchasable {
     private String name;
-    private int price;
+    private double price;
     private boolean available = true;
     private String description;
 
-    public Item(String name, int price, String description) {
+    public Item(String name, double price, String description) {
         this.name = name;
         this.price = price;
         this.description = description;
@@ -19,7 +19,7 @@ public abstract class Item implements Purchasable {
         return name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
